+++ src/lang.rs
@@ -25,16 +25,21 @@ pub enum Language {
     /// Polish
     #[serde(rename = "pl-PL")]
     Polish,
+
+    /// Brazilian Portuguese
+    #[serde(rename = "pt-BR")]
+    PortugueseBrazilian,
 }
 
 impl Language {
-    pub const ALL: &'static [Self] = &[Self::English, Self::French, Self::Polish];
+    pub const ALL: &'static [Self] = &[Self::English, Self::French, Self::Polish, Self::PortugueseBrazilian];
 
     pub fn id(&self) -> LanguageIdentifier {
         let id = match self {
             Self::English => "en-US",
             Self::French => "fr-FR",
             Self::Polish => "pl-PL",
+            Self::PortugueseBrazilian => "pt-BR",
         };
         id.parse().unwrap()
     }
@@ -44,6 +49,7 @@ impl Language {
             Self::English => "English",
             Self::French => "Français",
             Self::Polish => "Polski",
+            Self::PortugueseBrazilian => "Português brasileiro",
         }
     }
 
@@ -51,7 +57,8 @@ impl Language {
         match self {
             Self::English => 100,
             Self::French => 2,
-            Self::Polish => 84,
+            Self::Polish => 87,
+            Self::PortugueseBrazilian => 1,
         }
     }
 }
@@ -88,6 +95,7 @@ fn set_language(language: Language) {
         Language::English => include_str!("../lang/en-US.ftl"),
         Language::French => include_str!("../lang/fr-FR.ftl"),
         Language::Polish => include_str!("../lang/pl-PL.ftl"),
+        Language::PortugueseBrazilian => include_str!("../lang/pt-BR.ftl"),
     }
     .to_owned();